requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
motor>=3.1.1
pydantic[email]>=2.0.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
